Move relevant parts from PositionComponent to BaseComponent
